+++ plugin_ide.ui/src/com/github/rustdt/ide/ui/navigator/RustNavigatorActionProvider.java
@@ -12,17 +12,25 @@ package com.github.rustdt.ide.ui.navigator;
 
 import static melnorme.utilbox.core.CoreUtil.list;
 
+import java.nio.file.Path;
+
 import org.eclipse.core.resources.IProject;
+import org.eclipse.core.runtime.CoreException;
 import org.eclipse.jface.action.MenuManager;
 import org.eclipse.ui.IViewPart;
 
+import com.github.rustdt.ide.core.operations.RustSDKPreferences;
 import com.github.rustdt.ide.ui.launch.RustLaunchShortcut;
 
+import melnorme.lang.ide.core.operations.ILangOperationsListener_Default.ProcessStartKind;
+import melnorme.lang.ide.core.operations.ILangOperationsListener_Default.StartOperationOptions;
 import melnorme.lang.ide.ui.LangUIMessages_Actual;
 import melnorme.lang.ide.ui.launch.LangLaunchShortcut;
 import melnorme.lang.ide.ui.navigator.BuildTargetsActionGroup;
 import melnorme.lang.ide.ui.navigator.LangNavigatorActionProvider;
+import melnorme.lang.ide.ui.operations.ToolSourceModifyingOperation;
 import melnorme.lang.ide.ui.operations.RunToolOperation.RunSDKToolOperation;
+import melnorme.utilbox.core.CommonException;
 
 public class RustNavigatorActionProvider extends LangNavigatorActionProvider {
 	
@@ -39,18 +47,21 @@ public class RustNavigatorActionProvider extends LangNavigatorActionProvider {
 	@Override
 	protected void initActionGroups(IViewPart viewPart) {
 		super.initActionGroups(viewPart);
-		actionGroups.add(new DubPathActionGroup(viewPart));
+		actionGroups.add(new RustBundleOperationsActionGroup(viewPart));
 	}
 	
-	public static class DubPathActionGroup extends BundleOperationsActionGroup {
+	public static class RustBundleOperationsActionGroup extends BundleOperationsActionGroup {
 		
-		public DubPathActionGroup(IViewPart viewPart) {
+		public RustBundleOperationsActionGroup(IViewPart viewPart) {
 			super(viewPart);
 		}
 		
 		@Override
 		protected void initActions(MenuManager bundleOpsMenu, IProject project) {
-			addRunOperationAction(bundleOpsMenu, new AddDubProjectToLocalPath(project));
+			addRunOperationAction(bundleOpsMenu, 
+				new RunSDKToolOperation(LangUIMessages_Actual.CargoUpdate_OpName, project, list("update")));
+			addRunOperationAction(bundleOpsMenu, 
+				new CargoFormatOperation(project));
 		}
 		
 		@Override
@@ -58,11 +69,25 @@ public class RustNavigatorActionProvider extends LangNavigatorActionProvider {
 			return LangUIMessages_Actual.BundleOperationsMenu;
 		}
 		
-		public class AddDubProjectToLocalPath extends RunSDKToolOperation {
-			public AddDubProjectToLocalPath(IProject project) {
-				super(LangUIMessages_Actual.CargoUpdate_OpName, project,
-					list("update"));
-			}
+	}
+	
+	public static class CargoFormatOperation extends ToolSourceModifyingOperation {
+		public CargoFormatOperation(IProject project) {
+			super(
+				LangUIMessages_Actual.CargoFormat_OpName, 
+				project, 
+				list("--", "--write-mode=overwrite"), 
+				new StartOperationOptions(ProcessStartKind.BUILD, true, true)
+			);
+		}
+		
+		@Override
+		protected ProcessBuilder createProcessBuilder() throws CoreException, CommonException {
+			Path rustfmtPath = RustSDKPreferences.RUSTFMT_PATH.getDerivedValue();
+			// Look for sibling command cargo-fmt
+			Path cargoFmtPath = rustfmtPath.getParent().resolve("cargo-fmt");
+			
+			return getToolManager().createToolProcessBuilder(project, cargoFmtPath, getCommands());
 		}
 		
 	}